requirements.txt
# Core packages
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.1.0
scipy>=1.7.0

# Computer Vision and Embeddings  
open-clip-torch>=2.20.0
Pillow>=9.0.0
opencv-python>=4.6.0

# PyTorch (compatible versions)
torch>=1.12.0
torchvision>=0.13.0

# Visualization and UI
streamlit>=1.28.0
plotly>=5.11.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Dimensionality Reduction
umap-learn>=0.5.3

# Utilities
tqdm>=4.64.0

# Optional: Image Generation (for Stable Diffusion workflow)
# Install with: pip install diffusers accelerate transformers xformers
# diffusers>=0.21.0
# accelerate>=0.20.0
# transformers>=4.30.0
# xformers>=0.0.20